Refactor: Move router-related schemas to a new file and clean up RouterNode component
