+++ backend/app/nodes/logic/router.py
@@ -1,39 +1,9 @@
-from typing import Dict, Any, Optional, List, Literal
-from pydantic import BaseModel, Field, create_model
-from enum import Enum
+from typing import Dict, Any, Optional
+from pydantic import BaseModel, create_model
 
 from ..base import BaseNodeConfig, BaseNode, BaseNodeInput, BaseNodeOutput
 
-
-class ComparisonOperator(str, Enum):
-    CONTAINS = "contains"
-    EQUALS = "equals"
-    GREATER_THAN = "greater_than"
-    LESS_THAN = "less_than"
-    STARTS_WITH = "starts_with"
-    NOT_STARTS_WITH = "not_starts_with"
-    IS_EMPTY = "is_empty"
-    IS_NOT_EMPTY = "is_not_empty"
-    NUMBER_EQUALS = "number_equals"
-
-
-LogicalOperator = Literal["AND", "OR"]
-
-
-class Condition(BaseModel):
-    """Configuration for a single condition"""
-
-    variable: str
-    operator: ComparisonOperator = Field(default=ComparisonOperator.CONTAINS)
-    value: Any
-    logicalOperator: Optional[LogicalOperator] = Field(default="AND")
-
-
-class RouteCondition(BaseModel):
-    """Configuration for a route with multiple conditions"""
-
-    conditions: List[Condition]
-
+from ...schemas.router_schemas import ComparisonOperator, Condition, RouteCondition
 
 class RouterNodeConfig(BaseNodeConfig):
     """Configuration for the router node."""
@@ -170,7 +140,6 @@ class RouterNode(BaseNode):
 
 if __name__ == "__main__":
     # Test the RouterNode
-    from typing import List
     from pydantic import BaseModel
     import asyncio
